// apps/banana/types/viewport.types.ts
export enum VIEWPORT {
  mobile = 'mobile',
  tablet = 'tablet',
  desktop = 'desktop',
}

export enum BREAKPOINT {
  mobileXSmall = 320,
  mobileSmall = 350,
  mobileLarge = 450,
  tabletSmall = 600,
  tabletMedium = 768,
  desktopXSmall = 1023,
  desktopSmall = 1336,
  desktopMedium = 1600,
  desktopLarge = 1920,
  desktopXLarge = 2560,
}